local-backend/requirements.txt
flask
flask-cors
openai
pillow